Automatically backed up by Learn

+++ index.js
@@ -29,5 +29,5 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-  return parseInt(n);
+  return parseFloat(n);
 }
